added cookie and fine tuned some logic

+++ src/main/java/com/team66/controller/powerGeneratorController.java
@@ -15,13 +15,19 @@ public class powerGeneratorController {
     @Autowired
     private powerGeneratorService powerGeneratorService;
 
+    @GetMapping("/check-if-offthegrid")
+    public boolean checkIfOffTheGrid(@CookieValue(value = "cookieEmail") String email) {
+        return powerGeneratorService.checkIfOffTheGrid(email);
+    }
+
     @PostMapping("/add")
-    public void addPowerGenerator(@CookieValue(value = "cookieEmail") String email, @RequestBody Map<String, Object> queryMap) {
+    public String addPowerGenerator(@CookieValue(value = "cookieEmail") String email, @RequestBody Map<String, Object> queryMap) {
         Integer powerGeneratorNum = (Integer) queryMap.get("powerGeneratorNum");
         Integer monthlyKwh = (Integer) queryMap.get("monthlyKwh");
         Integer storageKwh = (Integer) queryMap.get("storageKwh");
         String generationType = (String) queryMap.get("generationType");
         powerGeneratorService.addPowerGenerator(new powerGenerator(email, powerGeneratorNum, monthlyKwh, storageKwh, generationType));
+        return "success";
     }
 
     @GetMapping("/view")
@@ -30,8 +36,12 @@ public class powerGeneratorController {
     }
 
     @DeleteMapping("/delete/{powerGeneratorNum}")
-    public List<Map<String, Object>> deletePowerGenerator(@CookieValue(value = "cookieEmail") String email, @PathVariable Integer powerGeneratorNum) {
-        powerGeneratorService.deletePowerGenerator(email, powerGeneratorNum);
-        return powerGeneratorService.viewPowerGenerator(email);
+    public String deletePowerGenerator(@CookieValue(value = "cookieEmail") String email, @PathVariable Integer powerGeneratorNum) {
+        if(checkIfOffTheGrid(email) && powerGeneratorService.viewPowerGenerator(email).size()==1)
+            return "not allowed";
+        else {
+            powerGeneratorService.deletePowerGenerator(email, powerGeneratorNum);
+            return "success";
+        }
     }
 }
